Gamming online almost implemented.

+++ src/org/academiadecodigo/bootcamp/filipejorge/Game.java
@@ -8,6 +8,7 @@ import org.academiadecodigo.simplegraphics.mouse.MouseEventType;
 import org.academiadecodigo.simplegraphics.mouse.MouseHandler;
 
 import java.awt.*;
+import java.io.IOException;
 
 /**
  * Created by filipejorge on 04/03/16.
@@ -35,7 +36,7 @@ public class Game {
     private Marker p1Marker;
     private Marker p2Marker;
 
-    private UDPConnection player2;
+    private UDPConnection player2Connection;
 
     private float fakeReceivedXPerCent = (float) (Math.random() * 100);
     private float fakeReceivedYPerCent = (float) (Math.random() * 100);
@@ -46,7 +47,6 @@ public class Game {
     }
 
     public void init() {
-
         field1 = new Field(factory.getGameObject(GameObjectType.FIELD, MARGIN_LEFT + 1, MARGIN_TOP + 1, SCREEN_WIDTH / 2, SCREEN_HEIGHT));
         field1.fill();
         field1.setColor(15, 15, 30);
@@ -54,6 +54,10 @@ public class Game {
         field2 = new Field(factory.getGameObject(GameObjectType.FIELD, MARGIN_LEFT + 1 + (SCREEN_WIDTH / 2), MARGIN_TOP + 1, SCREEN_WIDTH / 2, SCREEN_HEIGHT));
         field2.fill();
         field2.setColor(35, 15, 15);
+
+        System.out.println("Connecting");
+        player2Connection = new UDPConnection();
+        System.out.println("Connected");
 
         p1Marker = new Marker(factory.getGameObject(GameObjectType.MARKER, field1.getX(), field1.getY(), 30));
         p1Marker.delete();
@@ -64,7 +68,6 @@ public class Game {
         p2Marker.setColor(255, 210, 175);
 
         TaptapMouseHandler mouseHandler = new TaptapMouseHandler();
-
 
         start();
 
@@ -78,11 +81,16 @@ public class Game {
         //choose who starts randomly
         //fifty-fifty.
         if (Math.random() >= 0.5f) {
+            System.out.println("your turn");
             playerTurn = true;
-            p1Marker.setCenter(field1.getWidth() - field1.percentToX(fakeReceivedXPerCent), field1.getY() + field1.percentToY(fakeReceivedYPerCent));
-            p1Marker.fill();
+            markTapped = true;
+            //should go to phase2
+
+        } else {
+            System.out.println("other player turn");
         }
 
+        //game cycle
         while (true) {
             try {
                 Thread.sleep(1);
@@ -90,37 +98,48 @@ public class Game {
                 e.printStackTrace();
             }
 
-
-           /*
-           Marker temp = new Marker(factory.getGameObject(GameObjectType.MARKER, field1.getX(), field1.getY(), 30));
-            temp.setCenter(field1.percentToX(field2.xToPercent(field2.getRandomXInside(15))),field1.percentToY(field2.yToPercent(field2.getRandomYInside(15))));
-           */
-
-
             //receive x,y from udp
-                //fake, and if udp fails
+            //receives a string: "XXX.XX YYY.YY"
+            //should be transformed to 2 floats
+            //they are percentage values, so it can work with different screen sizes.
+
+            //fake, and if udp fails
+
+            //phase1 - player1 has to hit is marker, his marker position is sent by player2
+            //receiving data and setting up marker
             if (!playerTurn) {
-                System.out.println("received from p2");
+
+                String[] receivedString = new String[1];
+
+                try {
+                    receivedString = player2Connection.receiveDatagram().split(" ");
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+
+                float receivedXpercent = new Float(receivedString[0]);
+                float receivedYpercent = new Float(receivedString[1]);
+                System.out.println("received from p2 " + receivedXpercent + " " + receivedYpercent);
                 p2Marker.draw();
 
                 //receive a %, need to translate to this field x, y
-                p1Marker.setCenter(field1.getWidth() - field1.percentToX(fakeReceivedXPerCent), field1.getY() + field1.percentToY(fakeReceivedYPerCent));
-                //p1Marker.setPos(field1.getRandomXInside(p1Marker.getWidth()), field1.getRandomYInside(p1Marker.getHeight()));
+                p1Marker.setCenter(field1.getWidth() - field1.percentToX(receivedXpercent), field1.getY() + field1.percentToY(receivedYpercent));
                 p1Marker.fill();
 
                 playerTurn = true;
-                //start timer
+                //start timing for score
             }
 
             //detect if player1 hit marker
-            if (playerTurn && mouseClicked && (Math.hypot(p1Marker.getCenterX() - mouseX, p1Marker.getCenterY() - mouseY) <= p1Marker.getRadius())) {
+            if (playerTurn && mouseClicked && !markTapped &&
+                    (Math.hypot(p1Marker.getCenterX() - mouseX, p1Marker.getCenterY() - mouseY) <= p1Marker.getRadius())) {
                 markTapped = true;
                 mouseClicked = false;
                 p1Marker.delete();
                 System.out.println("Hit");
             }
 
-            //make player1 set player2 marker position
+            //phase 2 - player 1 sets the player2's marker position and send it
             if (playerTurn && mouseClicked && markTapped) {
                 mouseClicked = false;
                 // X, Y as to be in p2 Field
@@ -133,12 +152,19 @@ public class Game {
 
                     //sent by UDP to player2, as %
 
-                    fakeReceivedXPerCent = field2.xToPercent(p2X);
-                    fakeReceivedYPerCent = field2.yToPercent(p2Y);
-                    System.out.println("sent to p2: x " + field2.xToPercent(p2X) + " " + field2.yToPercent(p2Y));
+                    float sentXpercent = field2.xToPercent(p2X);
+                    float sentYpercent = field2.yToPercent(p2Y);
+
+                    try {
+                        player2Connection.sendDatagram(sentXpercent + " " + sentYpercent);
+                    } catch (IOException e) {
+                        //couldn't send
+                        e.printStackTrace();
+                    }
+                    System.out.println("sent to p2: x " + sentXpercent + " " + sentYpercent);
 
                     playerTurn = false;
-                    //stop timer, calc score
+                    //stop timing, calc score
                 }
 
             }
